1) Restructured the class MyBio::Data::File::SAM that serves as a file parser for a SAM formatted file. Significant improvements have been made and most of them break backwards compatibility. 2) Developed a class named MyBio::Data::File::SAM::Record that corresponds to a record stored in a SAM formatted file. 3) Developed unit tests for the above classes and cleaned up the t/ folder.
IMPORTANT:
From now on tests need to be created for any newly developed class. These tests should test EVERY method of the class and should be written in an object oriented way.
In the following commits we will try to develop as many tests as possible for the existent classes.