Modified IO design for JobGraph module
Summary:
The classes that handle the job input/output have been redesigned.
The reason is that the classes had to much control over the data object which they were
handling which made interconnection rather difficult. For example, having a file as IO,
the classes could directly make modifications to the files filename. This would create
sync problems when trying to convert an input object into an output object and vice versa
or when the same file would serve as input for one job and as output for another.