Test suite now loads test modules dynamically without Test::Class::Load
Summary:
The script previously used Test::Class::Load which resulted in warnings
at runtime which had to do with the dynamic loading of the test modules.
The use of this module was removed and a new implementation is provided.
It seems that the problem with the warnings has been resolved