Change the interface of Data::File::SAM::Record for optimization
Summary
The idea behind this series of optimizations is that in most cases when
a user loops over a SAM file he/she usually requires only a handful of
the exact SAM record fields and therefore the overhead of creating all
fields as attributes and calculating every single value on object instantiation
is to high. For this the SAM::Record accessors have been simplified to just a redirection
to the SAM::Record fields array (very fast) and everything else is calculated
from these fields on first request.