Fix gene creation from GTF
There was a bug in the gene GTF factory which resulted in the genes not being
created correctly. Specifically the number of created genes was semi-random
and the problem was coming from the definition of a gene based on transcripts
overlap. Some times a new gene was created by falsely identifying that its
transcripts did not overlpap with each other.