Skip genes whose transcripts do not overlap
In a GTF file there are cases in which the transcripts of a gene do not
overlap. Previously, a separate gene entry for each overlapping set of its
transcripts was created. We no longer allow this. It seems that these genes
are poorly annotated and therefore are now skipped entirely.